api/python/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
claude-code-sdk>=0.0.20
aiohttp>=3.9.0
